+++ src/vue-router/routes.js
@@ -12,4 +12,8 @@ export const vueRoutes = Object.freeze({
   // accounts group
   walletPage: { name: 'wallet' },
   myAccountsTab: { name: 'wallet.my-accounts-tab' },
+
+  // holders group
+  holdersPage: { name: 'holders' },
+
 })