+++ src/vue/composables/useMenuLinksGroups.js
@@ -28,6 +28,10 @@ export function useMenuLinksGroups () {
           translationName: t('routers-groups.temples.routers.event-calendar'),
           route: vueRoutes.eventCalendarPage,
         },
+        {
+          translationName: t('routers-groups.temples.routers.holders'),
+          route: vueRoutes.holdersPage,
+        },
       ],
     },
   ])